First DataBase refactoring
